Decouple between CI/CD and plugin sources
Before this change, a single source would include methods
related to built-in CI/CD systems/models and plugin models.

This change decouples between the two so if in the future, someone
who uses plugin "x" won't get the methods and related data of plugin
"y". Also, someone can use the app without any plugins and then
the sources won't be extended with plugin related methods.

The way it works:

1. For each plugin identify existing sources
2. Check if plugin source name matches one of the built-in sources
3 If yes, start extending built-in source with attributes from
  the matching plugin source
