+++ cibyl/plugins/__init__.py
@@ -13,18 +13,57 @@
 #    License for the specific language governing permissions and limitations
 #    under the License.
 """
+import inspect
 import logging
+import os
 
 from cibyl.exceptions.plugin import MissingPlugin
+from cibyl.sources.source_factory import SourceFactory
 
 LOG = logging.getLogger(__name__)
 
 
-def extend_models(plugin_name):
+def get_plugin_module(plugin_name: str):
+    """Returns plugin module based on a given plugin N
+
+       :param plugin_name: the name of the plugin
+       :type plugin_name: str
+    """
     try:
-        LOG.info("Loading plugin: %s", plugin_name)
-        loaded_plugin = __import__(f"cibyl.plugins.{plugin_name}",
-                                   fromlist=[''])
-        loaded_plugin.Plugin()._extend()
+        return __import__(f"cibyl.plugins.{plugin_name}",
+                          fromlist=[''])
     except (ImportError, ModuleNotFoundError):
-        raise MissingPlugin(plugin_name) from None
+        raise MissingPlugin(plugin_name)
+
+
+def get_plugin_module_path(plugin_module):
+    return os.path.join(os.path.dirname(os.path.abspath(
+        plugin_module.__file__)), 'sources')
+
+
+def is_plugin_class(class_obj):
+    return inspect.isclass(class_obj) and \
+            "sources" in class_obj.__module__
+
+
+def extend_source(plugin_name, plugin_module_path):
+    plugin_sources = []
+    for py in [f[:-3] for f in os.listdir(plugin_module_path)
+               if f.endswith('.py') and f != '__init__.py']:
+        mod = __import__('.'.join(
+            [f"cibyl.plugins.{plugin_name}.sources", py]), fromlist=[py])
+        plugin_sources = [source_tuple[1] for source_tuple in
+                          inspect.getmembers(mod, is_plugin_class)]
+    for plugin_source in plugin_sources:
+        SourceFactory.extend_source(plugin_source)
+
+
+def enable_plugins(plugins: list = None):
+    """Enable plugins based on a given list of plugins"""
+    if plugins:
+        for plugin in plugins:
+            LOG.debug("Loading plugin: %s", plugin)
+            plugin_module = get_plugin_module(plugin)
+            plugin_module.Plugin().extend_models()
+            plugin_module_path = get_plugin_module_path(plugin_module)
+            extend_source(plugin, plugin_module_path)
